requirements.txt
beautifulsoup4
requests
pytest
google-cloud-bigquery